machines: spider: Fixed a data abort exception in the CAN driver triggered by a 7-byte payload frame.
Receiving a CAN 2.0 frame with a 7-byte payload triggered a data abort exception.
This exception was due to the usage of memcpy() to copy the data from and to the
CAN controller registers.

Maybe the memcpy() has some internal optimization that copies 4 bytes at a time,
then single bytes, and these non uniform accesses were messing up the hardware.

Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>