spider: ethernet: clear L1 cache before reading RX descriptors
Without this fix it can happen that the descriptors read from CR52
are not aligned with the RAM content which results in no operation
on the received data. This is of course not correct and due to a
misalignment between the RAM content and L1 cache.
This commits resolves this issue by invalidating L1 cache before
reading descriptors.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>