goil: code: armv8: Fix ISR id mismatch
Currently, interrupt handler has same id so that interrupt doesn't work
correctly on envrionment which has multiple ISR.
This is caused by the scope of indexISR2 variable.

And, even if this issue is fixed, function order and index is not
matched always because function order is change by priotiry.
So, non-expected function may be called.

Thus, this patch fixes to use defined variable instead of index.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>